requirements.txt
# Meta Ads Agent Dependencies

# Meta/Facebook SDK
facebook-business>=19.0.0

# Claude API
anthropic>=0.18.0

# Dashboard
streamlit>=1.31.0
plotly>=5.18.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# HTTP & Utils
requests>=2.31.0
python-dotenv>=1.0.0

# Slack notification
slack-sdk>=3.27.0


